Re-organise --bundlers flag for tests
