+++ packages/core/src/constants.ts
@@ -3,3 +3,14 @@
 // Copyright 2019-Present Datadog, Inc.
 
 export const INJECTED_FILE = '__datadog-helper-file';
+
+export const ALL_BUNDLERS = ['webpack', 'vite', 'esbuild', 'rollup', 'rspack', 'rolldown', 'farm'];
+export const SUPPORTED_BUNDLERS = ['webpack', 'vite', 'esbuild', 'rollup', 'rspack'] as const;
+export const FULL_NAME_BUNDLERS = [
+    'webpack4',
+    'webpack5',
+    'vite',
+    'esbuild',
+    'rollup',
+    'rspack',
+] as const;
